fixed 【称重功能】购物车数量、商品库存等支持小数功能开发迭代

+++ fuint-application/src/main/java/com/fuint/common/dto/GoodsDetailDto.java
@@ -48,13 +48,13 @@ public class GoodsDetailDto implements Serializable {
     private BigDecimal linePrice;
 
     @ApiModelProperty("库存")
-    private Integer stock;
+    private Double stock;
 
     @ApiModelProperty("商品重量")
     private BigDecimal weight;
 
     @ApiModelProperty("初始销量")
-    private Integer initSale;
+    private Double initSale;
 
     @ApiModelProperty("商品卖点")
     private String salePoint;
