+++ fuint-application/src/main/java/com/fuint/common/dto/OrderDto.java
@@ -52,7 +52,7 @@ public class OrderDto implements Serializable {
     private Integer skuId;
 
     @ApiModelProperty("购买数量")
-    private Integer buyNum;
+    private Double buyNum;
 
     @ApiModelProperty("会员ID")
     private Integer userId;
