+++ fuint-repository/src/main/java/com/fuint/repository/model/MtGoods.java
@@ -63,7 +63,7 @@ public class MtGoods implements Serializable {
     private BigDecimal linePrice;
 
     @ApiModelProperty("库存")
-    private Integer stock;
+    private Double stock;
 
     @ApiModelProperty("关联卡券")
     private String couponIds;
@@ -75,7 +75,7 @@ public class MtGoods implements Serializable {
     private BigDecimal weight;
 
     @ApiModelProperty("初始销量")
-    private Integer initSale;
+    private Double initSale;
 
     @ApiModelProperty("商品卖点")
     private String salePoint;
